physical-ai-robotics/backend/requirements.txt
fastapi
uvicorn
openai-agents
python-dotenv
qdrant-client>=1.8.0
